Cgraphics_items splited in two files
