+++ CView/include/cgraphic_room.h
@@ -9,25 +9,9 @@ struct Meta {};
 struct Graph {};
 
 struct GraphNode {}; // структура, соответсвтующая комнате
-struct GraphConnection {}; // структура, соответствующая соединению
 
 namespace Polaris
 {
-class CGraphicConnection : public QGraphicsLineItem
-{
-public:
-    explicit CGraphicConnection( GraphConnection & connection_info, const QLineF & line,
-                                 QGraphicsItem * parent = nullptr );
-    size_t GetId();
-    double GetPrice();
-
-private:
-    size_t id_;
-    double price_;
-
-    void mousePressEvent( QMoveEvent * event );
-};
-
 class CGraphicRoom : public QGraphicsEllipseItem
 {
 public:
